% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_filter.R
\name{cor_filter}
\alias{cor_filter}
\title{Feature Selection Based on Pearson and Spearman Correlation}
\usage{
cor_filter(var_imp, data, threshold = NA_real_, pearson_threshold = 0.7,
  spearman_threshold = 0.7)
}
\arguments{
\item{var_imp}{Variables' importance given by \code{\link{var_imp}}}

\item{data}{the data set}

\item{threshold}{threshold for variable filtering on both Pearson and Spearman correlation, will overwrite \code{pearson_threshold} and \code{spearman_threshold}}

\item{pearson_threshold}{Threshold for variable filtering on Pearson correlation, will be neglected if \code{threshold} is set}

\item{spearman_threshold}{Threshold for variable filtering on Spearman correlation, will be neglected if \code{threshold} is set}
}
\value{
features selected
}
\description{
Feature Selection Based on Pearson and Spearman Correlation
}
