% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Variables' Importance}
\usage{
var_imp(model_tuned)
}
\arguments{
\item{model_tuned}{result given by function \code{\link{tune}}}
}
\value{
a data.table shown the importance of features
}
\description{
Showing the importance of variables in model.
}
